Profile Offline
Currently doesn’t update but shows everything it can offline (logout is
needed if you’ve been using the old version of the app)

+++ App/Memestagram/app/src/main/java/com/garethnunns/memestagram/LoginActivity.java
@@ -175,7 +175,7 @@ public class LoginActivity extends AppCompatActivity {
                                     loginEditor.putString("username",username);
                                     loginEditor.putString("password",password);
                                     loginEditor.putString("key",key);
-                                    loginEditor.putInt("iduser",id);
+                                    loginEditor.putLong("iduser",id);
 
                                     loginEditor.commit();
